+++ plugins/cost-insights/src/components/LegendItem/index.ts
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-export { LegendItem } from './LegendItem';
+export * from './LegendItem';
