+++ plugins/cost-insights/src/testUtils/public.ts
@@ -13,7 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export type Trendline = {
-  slope: number;
-  intercept: number;
-};
+
+export { MockConfigProvider, MockCurrencyProvider } from './providers';
