+++ plugins/cost-insights/src/types/index.ts
@@ -27,9 +27,8 @@ export * from './Filters';
 export * from './Group';
 export * from './Loading';
 export * from './Maybe';
-export * from './MetricData';
 export * from './Metric';
 export * from './Product';
 export * from './Project';
+export * from './ResourceData';
 export * from './Theme';
-export * from './Trendline';
